Split object TwitterSetupAndLog into two separate objects LoggerSetup and TwitterSetup for best design practice.
